+++ tests/unit/test_sfrFile_process.py
@@ -2,10 +2,10 @@ import pytest
 import requests
 
 from tests.helper import TestHelpers
-from processes import EpubProcess
+from processes import S3Process
 
 
-class TestEpubProcess:
+class TestS3Process:
     @classmethod
     def setup_class(cls):
         TestHelpers.setEnvVars()
@@ -16,27 +16,27 @@ class TestEpubProcess:
 
     @pytest.fixture
     def testInstance(self, mocker):
-        class TestEpubProcess(EpubProcess):
+        class TestS3Process(S3Process):
             def __init__(self, process, customFile, ingestPeriod):
                 self.bucket = 'testBucket'
         
-        return TestEpubProcess('TestProcess', 'testFile', 'testDate')
+        return TestS3Process('TestProcess', 'testFile', 'testDate')
 
     @pytest.fixture
     def testFile(self):
         return """
         {
-            "epubData": {
-                "epubURL": "testSourceURL",
+            "fileData": {
+                "fileURL": "testSourceURL",
                 "bucketPath": "testBucketPath"
             }
         }
         """
     
     def test_runProcess(self, testInstance, mocker):
-        mockReceive = mocker.patch.object(EpubProcess, 'receiveAndProcessMessages')
-        mockSave = mocker.patch.object(EpubProcess, 'saveRecords')
-        mockCommit = mocker.patch.object(EpubProcess, 'commitChanges')
+        mockReceive = mocker.patch.object(S3Process, 'receiveAndProcessMessages')
+        mockSave = mocker.patch.object(S3Process, 'saveRecords')
+        mockCommit = mocker.patch.object(S3Process, 'commitChanges')
 
         testInstance.runProcess()
 
@@ -45,8 +45,8 @@ class TestEpubProcess:
         mockCommit.assert_called_once
 
     def test_receiveAndProcessMessages(self, testInstance, mocker):
-        mockSleep = mocker.patch('processes.sfrEpub.sleep')
-        mockQueueGet = mocker.patch.object(EpubProcess, 'getMessageFromQueue')
+        mockSleep = mocker.patch('processes.s3Files.sleep')
+        mockQueueGet = mocker.patch.object(S3Process, 'getMessageFromQueue')
         mockProps = mocker.MagicMock()
         mockProps.delivery_tag = 'rabbitMQTag'
         mockQueueGet.side_effect = [
@@ -56,13 +56,13 @@ class TestEpubProcess:
             (None, None, None),
             (None, None, None)
         ]
-        mockStore = mocker.patch.object(EpubProcess, 'storeEpubFile')
-        mockAcknowledge = mocker.patch.object(EpubProcess, 'acknowledgeMessageProcessed')
+        mockStore = mocker.patch.object(S3Process, 'storeFileInS3')
+        mockAcknowledge = mocker.patch.object(S3Process, 'acknowledgeMessageProcessed')
 
         testInstance.receiveAndProcessMessages()
 
         assert mockQueueGet.call_count == 5
-        mockQueueGet.assert_called_with('test_epub_queue')
+        mockQueueGet.assert_called_with('test_file_queue')
 
         mockSleep.assert_has_calls([
             mocker.call(30), mocker.call(60), mocker.call(90)
@@ -71,12 +71,12 @@ class TestEpubProcess:
         mockStore.assert_called_once_with('epub_record')
         mockAcknowledge.assert_called_once_with('rabbitMQTag')
 
-    def test_storeEpubFile(self, testInstance, testFile, mocker):
-        mockGetContents = mocker.patch.object(EpubProcess, 'getFileContents')
+    def test_storeFileInS3(self, testInstance, testFile, mocker):
+        mockGetContents = mocker.patch.object(S3Process, 'getFileContents')
         mockGetContents.return_value = 'epubBinary'
-        mockPutInS3 = mocker.patch.object(EpubProcess, 'putObjectInBucket')
+        mockPutInS3 = mocker.patch.object(S3Process, 'putObjectInBucket')
 
-        testInstance.storeEpubFile(testFile)
+        testInstance.storeFileInS3(testFile)
 
         mockGetContents.assert_called_once_with('testSourceURL')
         mockPutInS3.assert_called_once_with('epubBinary', 'testBucketPath', 'testBucket')